[ 🔧 | #45 | Refactor ] 지역 랭크페이지 refactor
