+++ src/pages/RegionPage/RegionRankList/components/RegionRankDivider.jsx
@@ -1,0 +1,3 @@
+export default function RegionRankDivider() {
+	return <div className="my-auto h-48 w-[2px] rounded bg-greyBD" />;
+}
