+++ src/pages/RegionPage/RegionRankList/index.jsx
@@ -1,9 +1,10 @@
-import RegionRank1st from './RegionRank1st';
-import RegionRank2nd from './RegionRank2nd';
-import RegionRank3rd from './RegionRank3rd';
-import RegionRank4th from './RegionRank4th';
-import RegionRank5th from './RegionRank5th';
 import { regionRankData } from './_data/mock';
+import RegionRank1st from './components/RegionRank1st';
+import RegionRank2nd from './components/RegionRank2nd';
+import RegionRank3rd from './components/RegionRank3rd';
+import RegionRank4th from './components/RegionRank4th';
+import RegionRank5th from './components/RegionRank5th';
+import RegionRankDivider from './components/RegionRankDivider';
 
 function regionRankDataProps(data) {
 	return {
@@ -14,6 +15,14 @@ function regionRankDataProps(data) {
 		mvpId: data.mvpId,
 	};
 }
+
+const 랭크에따른컴포넌트 = {
+	레벨1: <RegionRank1st />,
+	레벨2: <RegionRank2nd />,
+	레벨3: <RegionRank3rd />,
+	레벨4: <RegionRank4th />,
+	레벨5: <RegionRank5th />,
+};
 
 export default function RegionRankList() {
 	const rankData = [];
@@ -26,91 +35,38 @@ export default function RegionRankList() {
 		<>
 			<div className="mt-8 ">
 				{/* 1열 */}
-				<div className=" mt-4 flex h-56 w-full flex-row rounded bg-white shadow-xl">
-					<div key={rankData[1].rank} className="w-full">
-						{/* 다른 필드들도 필요에 맞게 표시 */}
-						<RegionRank1st {...regionRankDataProps(rankData[1])} />
-					</div>
-				</div>
+				<RegionRank1st {...regionRankDataProps(rankData[1])} />
 
 				{/* 2열 */}
 				<div className="mt-4 flex h-56 w-full flex-row rounded  bg-white shadow-xl">
-					<div className=" w-1/2 ">
-						<div key={rankData[2].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank2nd {...regionRankDataProps(rankData[2])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className=" w-1/2 ">
-						<div key={rankData[3].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank2nd {...regionRankDataProps(rankData[3])} />
-						</div>
-					</div>
+					<RegionRank2nd {...regionRankDataProps(rankData[2])} />
+					<RegionRankDivider />
+					<RegionRank2nd {...regionRankDataProps(rankData[3])} />
 				</div>
 
 				{/* 3열 */}
 				<div className="mt-4 flex h-56 w-full flex-row rounded bg-white shadow-xl">
-					<div className="w-1/3">
-						<div key={rankData[4].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank3rd {...regionRankDataProps(rankData[4])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className="w-1/3">
-						<div key={rankData[5].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank3rd {...regionRankDataProps(rankData[5])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className="w-1/3">
-						<div key={rankData[6].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank3rd {...regionRankDataProps(rankData[6])} />
-						</div>
-					</div>
+					<RegionRank3rd {...regionRankDataProps(rankData[4])} />
+					<RegionRankDivider />
+					<RegionRank3rd {...regionRankDataProps(rankData[5])} />
+					<RegionRankDivider />
+					<RegionRank3rd {...regionRankDataProps(rankData[6])} />
 				</div>
 
 				{/* 4열 */}
 				<div className="mt-4 flex h-56 w-full flex-row rounded bg-white shadow-xl">
-					<div className="w-1/4">
-						<div key={rankData[7].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank4th {...regionRankDataProps(rankData[7])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className=" w-1/4">
-						<div key={rankData[8].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank4th {...regionRankDataProps(rankData[8])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className="w-1/4">
-						<div key={rankData[9].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank4th {...regionRankDataProps(rankData[9])} />
-						</div>
-					</div>
-					<div className="my-auto h-48 w-[2px] rounded bg-greyBD"></div>
-					<div className="w-1/4">
-						<div key={rankData[10].rank}>
-							{/* 다른 필드들도 필요에 맞게 표시 */}
-							<RegionRank4th {...regionRankDataProps(rankData[10])} />
-						</div>
-					</div>
+					<RegionRank4th {...regionRankDataProps(rankData[7])} />
+					<RegionRankDivider />
+					<RegionRank4th {...regionRankDataProps(rankData[8])} />
+					<RegionRankDivider />
+					<RegionRank4th {...regionRankDataProps(rankData[9])} />
+
+					<RegionRank4th {...regionRankDataProps(rankData[10])} />
 				</div>
 
 				{/* 5열 */}
 				<div className="mt-4 flex h-28 w-full flex-row rounded bg-white shadow-xl">
-					<div key={rankData[11].rank} className="w-full">
-						{/* 다른 필드들도 필요에 맞게 표시 */}
-						<RegionRank5th {...regionRankDataProps(rankData[11])} />
-					</div>
+					<RegionRank5th {...regionRankDataProps(rankData[11])} />
 				</div>
 			</div>
 		</>
